+++ src/main/java/com/example/oneplusone/domain/common/redisLock/RedissonLockService.java
@@ -1,4 +1,4 @@
-package com.example.oneplusone.domain.common.redis;
+package com.example.oneplusone.domain.common.redisLock;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
