+++ src/test/java/com/example/oneplusone/domain/product/service/InitProductsTest.java
@@ -24,13 +24,13 @@ public class InitProductsTest {
     @DisplayName("5만개의 더미 데이터 생성")
     void createProducts() {
         User user = new User("admin", "admin", "password", UserRole.SELLER);
-        userRepository.save(user);
+        User saveUser = userRepository.save(user);
         List<Product> productList = new ArrayList<>();
         for(int i = 1; i <= 50000; i ++){
             String name = "product" + i;
             Long price = 10000L + i;
             Long quantity = 1L + i;
-            Product product = new Product(name, "type", price, quantity, user);
+            Product product = new Product(name, "type", price, quantity, saveUser);
             productList.add(product);
         }
         productRepository.saveAll(productList);
